[#56] refactor: comment-item 컴포넌트 분리 및 api 호출 방법 롤백
